refactor: component-protocol merge identity.proto into protocol.proto
