+++ providers/component-protocol/protobuf/proto/cp/protocol.proto
@@ -5,7 +5,7 @@ package erda.cp;
 import "google/api/annotations.proto";
 import "google/protobuf/struct.proto";
 
-option go_package = "github.com/erda-project/erda-infra/provider/component-protocol/protobuf/proto-go/cp/pb";
+option go_package = "github.com/erda-project/erda-infra/providers/component-protocol/protobuf/proto-go/cp/pb";
 
 service CPService {
   rpc Render (RenderRequest) returns (RenderResponse) {
@@ -126,3 +126,19 @@ message RenderResponse {
   ComponentProtocol protocol = 2;
 }
 
+// IdentityInfo
+message IdentityInfo {
+  // UserID is user id. It must be provided in some cases.
+  // Cannot be null if InternalClient is null.
+  // +optional
+  string userID = 1;
+
+  // InternalClient records the internal client, such as: bundle.
+  // Cannot be null if UserID is null.
+  // +optional
+  string internalClient = 2;
+
+  // OrgID is org id. It must be provided in some cases.
+  // +optional
+  string orgID = 3;
+}
